Move shared functions to `mod rpc`
