+++ crates/sncast/src/lib.rs
@@ -31,7 +31,8 @@ use starknet::{
 };
 
 use crate::response::print::print_as_warning;
-use shared::{consts::EXPECTED_RPC_VERSION, get_rpc_version, is_supported_version};
+use shared::consts::EXPECTED_RPC_VERSION;
+use shared::rpc::{get_rpc_version, is_supported_version};
 use std::collections::HashMap;
 use std::thread::sleep;
 use std::time::Duration;
